requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
tqdm>=4.66.4
pymongo[srv]>=4.6.0
Flask>=3.0.0
